requirements.txt
asgiref==3.2.3
attrs==19.3.0
autobahn==19.10.1
Automat==0.8.0
certifi==2019.9.11
cffi==1.13.2
channels==2.3.1
chardet==3.0.4
constantly==15.1.0
cryptography==2.8
daphne==2.3.0
Django==2.2.7
django-channels==0.7.0
django-crispy-forms==1.8.0
django-ipware==2.1.0
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
oauthlib==3.1.0
pycparser==2.19
PyHamcrest==1.9.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
sqlparse==0.3.0
Twisted==19.7.0
txaio==18.8.1
urllib3==1.25.6
zope.interface==4.6.0
